Move isListItemStructurallyEmpty to lexical_helper
